requirements.txt
numpy
pandas
sklearn
matplotlib
Django==3.0.3
jupyter
finnhub
djangorestframework